Move Pos2, Vel2 and Accel2 to physics2d package
